Move utilities into a sub-module.
This divides common things from the worker utilities and the gui utilities.
