+++ lyse/figure_manager.py
@@ -16,6 +16,8 @@ import lyse
 from collections import OrderedDict
 import sys
 
+from labscript_utils import dedent
+
 class FigureManager(object):
 
     def __init__(self):
@@ -69,7 +71,7 @@ class FigureManager(object):
                     To permanently fix this, please ensure you call figure() prior
                     to other matplotlib plotting functions.
                     """
-                sys.stderr.write(lyse.dedent(msg))
+                sys.stderr.write(dedent(msg))
             self.figs[identifier] = fig
             self.__allocated_figures.append(identifier)
             self._remove_dead_references(identifier, fig)
