+++ lyse/utils/tempfile2clipboard.py
